Cat API Integrated, Contect created design UI changes
