+++ src/Pages/Home/home.css
@@ -32,7 +32,7 @@
 
 .api-list-blocks .api-block .block-footer span {
   font-size: 14px;
-  font-weight: 600;
+  font-weight: 500;
   color: rgb(18, 75, 94);
 }
